% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{comp.prep}
\alias{comp.prep}
\title{Preparations to fit a genetic-spatial competition model}
\usage{
comp.prep(
  data,
  gen,
  repl,
  row,
  col,
  ind,
  trait,
  dist.row,
  dist.col,
  method,
  area = NULL,
  age = NULL,
  n.dec = 2,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the phenotypic data.}

\item{gen, repl, row, col, ind, trait}{A string. The name of the columns that correspond
to the genotype, replicate, row, column, individual and trait information,
respectively. \code{ind} can also distinguish the plots, if there are multiple plants per plot.}

\item{dist.row, dist.col}{An integer. The distance between rows and columns in the trial.}

\item{method}{A string. The method for computing the competition intensity in \eqn{\mathbf{Z}_c}.
It has three options: "MU" for the method proposed by \insertCite{muir_incorporation_2005;textual}{competition},
"CC" for the method proposed by \insertCite{cappa_direct_2008;textual}{competition}, and."SK" for the
method proposed by \insertCite{costa_e_silva_accounting_2013;textual}{competition}.
See Details for more information on these methods.}

\item{area}{A string. The name of the column that corresponds to the area information.
Valid if the trial has non-contiguous blocks, for e.g., blocks 1 and 2 in area 1, and
blocks 3 and 4 in area 2. \code{NULL} (default) otherwise.}

\item{age}{A string. The name of the column that corresponds to the age information.
Necessary for fitting a multi-age model using \code{\link[=comp.asr]{comp.asr()}}. \code{NULL} (default)
otherwise.}

\item{n.dec}{An integer. The number of decimal digits to be displayed in \eqn{\mathbf{Z}_c}.
Defaults to 2.}

\item{verbose}{A logical value. If \code{TRUE}, a progress bar will be displayed in the
console. Defaults to \code{FALSE}.}
}
\value{
The function returns:
\itemize{
\item \code{Z} : The \eqn{\mathbf{Z}_c} matrix
\item \code{data} : A data frame composed of the built \eqn{\mathbf{Z}_c} merged.
with the dataset provided by the user.
\item \code{neigh_check} : A data frame containing the phenotypic records of each
focal plant and its neighbours.
\item \code{control} : A data frame containing the number of phenotypic records,
genotypes, replicates, rows, columns, ages and areas in the dataset provided by the user.
}
If \code{age} is not \code{NULL}, the function provides the outputs described for each age.
}
\description{
This function builds the genetic competition matrix (\eqn{\mathbf{Z}_c}), prepare the
dataset to be used to fit the model and provide summaries of the trial. It also
computes the competition intensity factor.
}
\details{
Three methods are available for estimating the competition intensity and building
the \eqn{\mathbf{Z}_c}, the genetic competition matrix:

\itemize{\item \insertCite{muir_incorporation_2005;textual}{competition}: "MU"}

The average competition intensity is the inverse of the distance between the focal
plant and its neighbours:

\deqn{f_{D_v} = \frac{1}{\sqrt{d_R^2 + d_C^2}}}

\deqn{f_{R_v} = \frac{1}{d_R}}

\deqn{f_{C_v} = \frac{1}{d_C}}

where \eqn{f_{D_v}}, \eqn{f_{R_v}} and \eqn{f_{C_v}} are the average competition intensities
in the diagonal, row, and column directions of the \eqn{v^{th}} clone,
respectively; and \eqn{d_R} and \eqn{d_C} are the inter-row and inter-column distances.

\itemize{\item \insertCite{cappa_direct_2008;textual}{competition}: "CC"}

The average compentition intensity depends on the number of neighbours in each direction

\deqn{f_{D_v} = \frac{1}{\sqrt{2(n_{C_v} + n_{R_v}) + n_{D_v}}}}

\deqn{f_{R_v} = \sqrt{\frac{2}{2(n_{C_v} + n_{R_v}) + n_{D_v}}}}

\deqn{f_{C_v} = \sqrt{\frac{2}{2(n_{C_v} + n_{R_v}) + n_{D_v}}}}

where \eqn{n_{D_v}}, \eqn{n_{R_v}} and \eqn{n_{C_v}} are the number of neighbours in the
diagonal, row and column directions of the \eqn{v^{th}} clone, respectively.
Note that, in this case, it is assumed that the distance between rows and
columns are the same.

\itemize{\item \insertCite{costa_e_silva_accounting_2013;textual}{competition}: "SK"}

The average competition intensity depends on both the distance between the focal
tree and its neighbours, and the number of neighbours in each direction:

\deqn{f_{D_v} = \frac{p}{\sqrt{(n_{R_v} p^4) + (n_{R_v} p^2) + (n_{C_v} p^2) + (n_{D_v} p^2) + n_{C_v}}}}

\deqn{f_{R_v} = f_{D_v} \sqrt{1 + p^2}}

\deqn{f_{C_v} = \frac{f_{D_v} \sqrt{1 + p^2}}{p}}

where \eqn{p = \frac{d_C}{d_R}}.

The overall competition intensity factor (\eqn{CIF}) is estimated by taking the mean of
the competition intensities provided by the equations described above, and the
mean number of neighbours in each direction:

\deqn{CIF = \overline{n_D} \overline{f_D} + \overline{n_R} \overline{f_R} + \overline{n_C} \overline{f_C}}
}
\examples{
\donttest{
 comp_mat = comp.prep(data = eucalyptus, gen = 'clone', repl = 'block', area = 'area', 
                      ind = 'tree', age = 'age', row = 'row', col = 'col', 
                      dist.col = 3, dist.row = 2, trait = 'mai', method = 'SK',
                      n.dec = 3, verbose = TRUE)
}
}
\references{
\insertAllCited{}
}
