% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition-package.R
\docType{package}
\name{competition-package}
\alias{competition}
\alias{competition-package}
\title{competition: Fit Spatial-Genetic Competition Models}
\description{
Fit spatial-genetic competition models for assessing the competition capacity of genotypes. Contained are functions used to build a competition matrix based on different criteria, fit the spatial-genetic competition model, and extract and visualize the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kaio-Olimpio/Competition}
  \item Report bugs at \url{https://github.com/Kaio-Olimpio/Competition/issues}
}

}
\author{
\strong{Maintainer}: Saulo Chaves \email{saulo.chaves@ufv.br} (\href{https://orcid.org/0000-0002-0694-1798}{ORCID})

Authors:
\itemize{
  \item Kaio Dias \email{kaio.o.dias@ufv.br} (\href{https://orcid.org/0000-0002-9171-1021}{ORCID}) [copyright holder]
  \item Filipe Ferreira \email{ferreira.fmanoel@gmail.com} (\href{https://orcid.org/0000-0002-7847-8333}{ORCID})
}

}
\keyword{internal}
