% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{.checkped}
\alias{.checkped}
\title{Check a pedigree (adapted from the \code{asreml} package)}
\usage{
.checkped(pedigree, mv = c("0", "NA"))
}
\arguments{
\item{pedigree}{A data frame where the first three columns correspond to the identifiers for the individual, male parent and female parent, respectively. Founders or unknown parents use 0 (zero) or NA in the parental columns.}

\item{mv}{A character vector of missing value indicators; elements of pedigree that exactly match any of the members of mv are treated as missing. The values passed should be used in addition to "0" and "NA" (default = "0", "NA").}
}
\value{
The pedigree in a data frame, reordered and expanded if necessary.
}
\description{
Checks a pedigree for consistency, adds missing founders and sorts so that founders appear before individuals.
Adapted from \code{asreml} V. 4.2.0.481
}
\keyword{internal}
