% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{.up.mod}
\alias{.up.mod}
\title{Update the asreml object}
\usage{
.up.mod(model)
}
\arguments{
\item{model}{An object of class \code{asreml}.}
}
\value{
An object of class \code{asreml} with all parameters converged.
}
\description{
This function keeps updating the \code{asreml} object until full convergence
}
\seealso{
\code{\link[=asr]{asr()}}, \code{\link[=asr_ma]{asr_ma()}},  \link[asreml:asreml.options]{asreml::asreml.options}, \link[asreml:asreml.object]{asreml::asreml.object}, \link[asreml:family_dist]{asreml::family_dist}
}
\keyword{internal}
