% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{plot.comprepfor}
\alias{plot.comprepfor}
\title{Plots for data overview}
\usage{
\method{plot}{comprepfor}(x, ..., category = "heatmap", age = "all")
}
\arguments{
\item{x}{An object of class \code{comprepfor}}

\item{...}{Currently not used.}

\item{category}{A string indicating which plot to build. Options are \code{"heatmap"} for
plotting the field grid (default), or \code{"boxplot"} for plotting the boxplots of phenotypic
performance.}

\item{age}{A string indicating if plots should be built per age. Options are
\code{"all"} for plotting all ages using facets, or the name of the specific age. Defaults
to \code{"all"}, which also serves when data has a single age}
}
\description{
The function generates two types of plots: i) a heatmap representing the grid.
The cells are filled according to the phenotype value of each plot; and ii)
boxplots depicting the performance of each selection candidate
}
\examples{
\donttest{
library(gencomp)
 comp_mat = prepfor(data = euca, gen = 'clone', area = 'area',
                    plt = 'tree', age = 'age', row = 'row', col = 'col',
                    dist.col = 3, dist.row = 2, trait = 'MAI', method = 'SK',
                    n.dec = 3, verbose = FALSE, effs = c("block"))
                                     
plot(comp_mat, category = 'heatmap', age = "all")
plot(comp_mat, category = 'boxplot', age = "3y")
# Note that the ages are labelled as "3y" and "6y" in the example dataset 
}

}
\seealso{
\link[ggplot2:ggplot2]{ggplot2}, \link{prepfor}
}
