% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato.R
\docType{data}
\name{potato}
\alias{potato}
\title{Potato dataset}
\format{
\subsection{\code{potato}}{

A data frame with 80 rows and 6 columns:
\describe{
\item{rep}{4 repetitions}
\item{gen}{20 genotypes}
\item{row}{4 row}
\item{col}{20 columns}
\item{yield}{80 observations}
\item{matur}{3 maturity classes}
}
}
}
\usage{
potato
}
\description{
This dataset is publicly available at the \code{agridat} package as \code{agridat::conolly.potato}.
Originally, it was used by Connolly et al. (1993) to study inter-plot competition
in single-drill plot trials in potato. They measured the tuber yield of 20 varieties
replicated four times. Each replication was an independent row of 20 drills.
The maturity class of each variety is also registered.
}
\references{
Connolly, T., Currie I.D., Bradshaw, J.E., McNicol, J.W. 1993. Inter-plot competition
in yield trials of potatoes Solanum tuberosum L. with single-drill plots.
Annals of Applied Biology, 123, 367-377. \doi{10.1111/j.1744-7348.1993.tb04099.x}
}
\keyword{dataset}
