% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepcrop}
\alias{prepcrop}
\title{Preparations to fit a genetic competition model for agronomic data}
\usage{
prepcrop(
  data,
  gen,
  row,
  col,
  plt,
  trait,
  effs = NULL,
  direction = "row",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the phenotypic data.}

\item{gen, row, col, plt, trait}{A string. The name of the columns that correspond
to the genotype, row, column, plot and trait information, respectively.}

\item{effs}{a string vector with column names of other effects that will be
considered in model fitting step. Defaults to \code{NULL} (if there is no further effect).}

\item{direction}{A string. The direction considered for building the competition matrix.
Can be \code{row} or \code{column}.}

\item{verbose}{A logical value. If \code{TRUE}, a progress bar will be displayed in the
console. Defaults to \code{FALSE}.}
}
\value{
The function returns:
\itemize{
\item \code{Z} : The competition incidence matrix (\eqn{\mathbf{Z}_c}).
\item \code{data} : A data frame composed of the built \eqn{\mathbf{Z}_c} merged
with the dataset provided by the user.
\item \code{neigh_check} : A data frame containing the phenotypic records of each
focal plant and its neighbours.
}
}
\description{
This function builds the genetic competition matrix, represented by a within-row
or within-column incidence matrix (Stringer et al., 2011). It also
prepares the dataset to be used for model fitting.
}
\references{
Stringer, J.K., Cullis, B.R., Thompson, R., 2008. Joint modeling of spatial
variability and within-row interplot competition to increase the efficiency of plant improvement.
Journal of Agricultural, Biological, and Environmental Statistics 16(2), 269-281. \doi{10.1007/s13253-010-0051-5}
}
