% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{print.comprepcrop}
\alias{print.comprepcrop}
\title{Print an object of class \code{comprepcrop}}
\usage{
\method{print}{comprepcrop}(x, ..., category = "matrix")
}
\arguments{
\item{x}{An object of class \code{comprepcrop}}

\item{...}{Currently not used.}

\item{category}{A string indicating which object to print. Options are \code{"all"} for
printing all objects, \code{"data"} for printing the data that will be used in the model,
\code{"matrix"} for printing the competition matrix, and \code{"check"} for printing the \code{neigh_check}
dataframe.}
}
\description{
Print a \code{comprepcrop} object in the R console.
}
\seealso{
\link{prepcrop}
}
