% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{print.comprepfor}
\alias{print.comprepfor}
\title{Print an object of class \code{comprepfor}}
\usage{
\method{print}{comprepfor}(x, ..., category = "matrix", age = "all")
}
\arguments{
\item{x}{An object of class \code{comprepfor}}

\item{...}{Currently not used.}

\item{category}{A string indicating which object to print. Options are \code{"all"} for
printing all objects, \code{"data"} for printing the data that will be used in the model,
\code{"matrix"} for printing the competition matrix, \code{"check"} for printing the \code{neigh_check}
dataframe, and \code{"CIF"} for printing the competition intensity factor.}

\item{age}{A string indicating if objects should be printed per age. Options
are \code{"all"} for printing all ages, or the name of the specific age. Defaults to
\code{"all"}, which also serves when data has a single age.}
}
\description{
Print a \code{comprepfor} object in the R console.
}
\seealso{
\link{prepfor}
}
