% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{print.comresp}
\alias{print.comresp}
\title{Print an object of class \code{comresp}}
\usage{
\method{print}{comresp}(x, ..., category = "blup.main", age = "all")
}
\arguments{
\item{x}{An object of class \code{comresp}.}

\item{...}{Currently not used.}

\item{category}{A string indicating which object to print. Options are \code{"all"}
for printing all objects, \code{"summar"} for printing the variance components and the
likelihood ratio test (if \code{lrt = TRUE} in \link{asr}),
\code{"blup.main"} (Default) for printing the DGE, IGE and TGV (the main effects
if a multi-age model was fitted), and \code{"blup.within"} for printing the DGE,
IGE and TGV within ages (if a multi-age model was fitted).}

\item{age}{If \code{category = 'blup.within'}, a string indicating if this object should be printed per age.
Options are \code{"all"} for printing all ages, or the name of a specific age. Defaults
to \code{"all"}.}
}
\description{
Print a \code{comresp} object in the R console.
}
\examples{
\donttest{
 library(gencomp)
  comp_mat = prepfor(data = euca, gen = 'clone', area = 'area',
                   plt = 'tree', age = 'age', row = 'row', col = 'col',
                   dist.col = 3, dist.row = 2, trait = 'MAI', method = 'SK',
                   n.dec = 3, verbose = FALSE, effs = c("block"))
 mod = asr(prep.out = comp_mat,
             fixed = MAI ~ age, 
             random = ~ block:age, 
             lrtest = TRUE, 
             spatial = TRUE, 
             cor = TRUE, 
             K = NULL,
             maxit = 20)
             
 results = resp(prep.out = comp_mat, model = mod, weight.tgv = FALSE, sd.class = 1)
 
 print(results)
 print(results, category = 'summar')
 print(results, category = 'blup.main')
 print(results, category = 'blup.within')
 print(results, category = 'blup.within', age = '6y')
 # Note that the ages are labelled as "3y" and "6y" in the example dataset 
 }
}
\seealso{
\link{resp}
}
