% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{summary.comprepcrop}
\alias{summary.comprepcrop}
\title{Summary of the \code{comprepcrop} object}
\usage{
\method{summary}{comprepcrop}(object, ...)
}
\arguments{
\item{object}{An object of class \code{comprepcrop}}

\item{...}{Currently not used.}
}
\description{
A brief summary of the dataset used.
}
\seealso{
\link{prepcrop}
}
