% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{summary.comprepfor}
\alias{summary.comprepfor}
\title{Summary of the \code{comprepfor} object}
\usage{
\method{summary}{comprepfor}(object, ...)
}
\arguments{
\item{object}{An object of class \code{comprepfor}}

\item{...}{Currently not used.}
}
\description{
A brief summary of the dataset used.
}
\seealso{
\link{prepfor}
}
