% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite}
\alias{composite}
\title{@importFrom stats quantile model.matrix model.matrix.lm

@export}
\usage{
composite(
  prep.out,
  model,
  resp.out,
  d.row.col,
  d.weight = TRUE,
  selected,
  nsim = 10,
  verbose = TRUE
)
}
\description{
@importFrom stats quantile model.matrix model.matrix.lm

@export
}
