% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencomp-package.R
\docType{package}
\name{gencomp-package}
\alias{gencomp}
\alias{gencomp-package}
\title{gencomp: Fit Spatial-Genetic Competition Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fit spatial-genetic competition models for assessing the competition capacity of genotypes. Contained are functions used to build a competition matrix based on different criteria, fit the spatial-genetic competition model, and extract and visualize the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kaio-Olimpio/gencomp}
  \item \url{https://kaio-olimpio.github.io/gencomp/}
  \item Report bugs at \url{https://github.com/Kaio-Olimpio/gencomp/issues}
}

}
\author{
\strong{Maintainer}: Saulo F. S. Chaves \email{saulochaves@usp.br} (\href{https://orcid.org/0000-0002-0694-1798}{ORCID})

Authors:
\itemize{
  \item Kaio Olimpio G. Dias \email{kaio.o.dias@ufv.br} (\href{https://orcid.org/0000-0002-9171-1021}{ORCID}) [copyright holder]
  \item Filipe M. Ferreira \email{ferreira.fmanoel@gmail.com} (\href{https://orcid.org/0000-0002-7847-8333}{ORCID})
  \item Getúlio C. Ferreira \email{getuliocaifer@gmail.com} (\href{https://orcid.org/0000-0003-4982-8274}{ORCID})
  \item Salvador A. Gezan \email{forestats.sg@gmail.com} (\href{https://orcid.org/0000-0002-5025-2658}{ORCID})
}

}
\keyword{internal}
