% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{plot.comprepcrop}
\alias{plot.comprepcrop}
\title{Plots for data overview}
\usage{
\method{plot}{comprepcrop}(x, ..., category = "heatmap")
}
\arguments{
\item{x}{An object of class \code{comprepcrop}}

\item{...}{Currently not used.}

\item{category}{A string indicating which plot to build. Options are \code{"heatmap"} for
plotting the field grid (default), or \code{"boxplot"} for plotting the boxplots of phenotypic
performance.}
}
\description{
The function generates two types of plots: i) a heatmap representing the grid.
The cells are filled according to the phenotype value of each plot; and ii)
boxplots depicting the performance of each selection candidate
}
\seealso{
\link{ggplot2}, \link{prepcrop}
}
