% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{plot.comresp}
\alias{plot.comresp}
\title{Plots for the \code{comresp} object}
\usage{
\method{plot}{comresp}(object, category = "DGE.IGE", level = "main", age = "all", ...)
}
\arguments{
\item{object}{An object of class \code{comresp}.}

\item{category}{A string indicating which plot to build. See options in the Details section.}

\item{level}{A string indicating the information level to be used for building
the plots. Options are \code{"main"} for focusing on the main effects, or \code{"within"} to
focus on the within-age effects effects (main + interaction effects). Defaults
to \code{"main"}, which also serves when data has a single age.}

\item{age}{If \code{level = 'within'}, a string indicating if plots should be built per age.
Options are \code{"all"} for plotting all ages, or the name of the specific age. Defaults
to \code{"all"}, which serves when data has a single age.}

\item{...}{Currently not used.}
}
\description{
Build plots using the outputs stored in the \code{comresp} object.
}
\details{
The available options are:
\itemize{
\item \code{class}: a density plot with the IGE distribution. The area within the
distribution is filled according to the competition class.
\item \code{DGEvIGE}: a scatter plot illustrating the relation between IGE (\emph{x}-axis)
and DGE (\emph{y}-axis). The dots are coloured according to the competition class
\item \code{DGE.IGE}: lollipop plots representing the DGE and IGE of each genotype. The plots
are in descending order according to the DGE. The dots' colour depicts the DGE's and IGE's
reliability of each genotype.
\item \code{TGV}: lollipop plots with the TGV of each genotype, in increasing order.
\item \code{nneigh}: a bar plot depicting the number of different genotypes as neighbours
(total and per competition class) of each selection candidate.
\item \code{grid.res}: a heatmap representing the grid. The cells are filled according
to the residual value of each plot.
\item \code{grid.dge}: a heatmap representing the grid. The cells are filled according
to the DGE value of each genotype.
\item \code{grid.ige}: a heatmap representing the grid. The cells are filled according
to the IGE value of each genotype.
\item \code{grid.class}: a heatmap representing the grid. The cells are filled according
to the competition class of each genotype.
}
}
\examples{
\donttest{
 library(GenComp)
 comp_mat = prep(data = euca, gen = 'clone', repl = 'block', area = 'area', 
                 ind = 'tree', age = 'age', row = 'row', col = 'col', 
                 dist.col = 3, dist.row = 2, trait = 'mai', method = 'SK',
                 n.dec = 3, verbose = TRUE)
 model = asr(prep.out = comp_mat, 
             fixed = mai~ age, 
             random = ~ block:age, 
             cor = TRUE, maxit = 20,
             lrtest = FALSE)
 results = resp(prep.out = comp_mat, model = model, weight.tgv = FALSE)
 
 plot(results, category = 'DGEvIGE', level = 'within', age = '6y')
 plot(results, category = 'grid.res', level = 'main', age = '3y')
 plot(results, category = 'nneigh', level = 'main')
 plot(results, category = 'DGE.IGE', level = 'within', age ='all')
 plot(results, category = 'grid.dge', level = 'within', age ='3y')
 plot(results, category = 'grid.ige', level = 'within', age ='6y')
 plot(results, category = 'grid.class', level = 'within', age ='all')
 plot(results, category = 'class', level = 'main')
 plot(results, category = 'TGV', level = 'within')
 # Note that the ages are labelled as "3y" and "6y" in the example dataset 
 }
}
\seealso{
\link{ggplot2}, \link{resp}
}
