% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{resp}
\alias{resp}
\title{Extract outputs from the genetic-spatial competition model}
\usage{
resp(prep.out, model, weight.tgv = FALSE)
}
\arguments{
\item{prep.out}{A \code{prep} object.}

\item{model}{An \code{asreml} object, preferably obtained using the \code{\link[=asr]{asr()}} function.}

\item{weight.tgv}{A logical value. If \code{TRUE}, the function will use the direct and
indirect genetic effects' reliability as a weight when estimating the total genotypic
value. Defaults to \code{FALSE}. See Details for more information on these methods.}
}
\value{
The function returns:
\itemize{
\item \code{lrt} : a data frame with the Likelihood Ratio Test results.
\item \code{varcomp} : A data frame summarizing the random parameter vector
(object$vparameters). Variance component ratios are included if param = "gamma",
and a measure of precision (standard error) is included along with boundary
constraints at termination and the percentage change in the final iteration.
\item \code{blups} : The direct (DGE) and indirect genetic effects (IGE), their standard errors,
the competition class of each genotype and the total genotypic value (TGV). If \code{age = TRUE},
\code{blup} is a list containing the main effects and the within-ages DGE and IGE. If
other random effects were declared in the model, \code{blup} will contain another data frame
with their BLUPs.
\item \code{plots} : A list of plots:
\itemize{
\item \code{IGE.density}: a density plot with the IGE distribution. The area within the
distribution is filled according to the competition class (see Details).
\item \code{DGEvsIGE}: a scatter plot illustrating the relation between IGE (\emph{x}-axis)
and DGE (\emph{y}-axis). The dots are coloured according to the competition class
\item \code{DGE.IGE}: a lollipop plots representing the DGE and IGE of each genotype. The plots
are in descending order according to the DGE. The dots' colour depicts the DGE's and IGE's
reliability of each genotype.
\item \code{TGV}: lollipop plots with the TGV of each genotype, in increasing order.
\item \code{n.neigh}: a bar plot depicting the number of different genotypes as neighbours
(total and per competition class) of each selection candidate. If \code{age = TRUE},
there will be a \code{n.neigh} plot for each age.
\item \code{grid.pheno}: a heatmap representing the grid. The cells are filled according
to the phenotype value of each plot. If \code{age = TRUE}, there will be a \code{grid.pheno} plot for each age.
\item \code{grid.DGE}: a heatmap representing the grid. The cells are filled according
to the DGE value of each genotype. If \code{age = TRUE}, there will be a \code{grid.DGE} plot for each age.
\item \code{grid.IGE}: a heatmap representing the grid. The cells are filled according
to the IGE value of each genotype. If \code{age = TRUE}, there will be a \code{grid.DGE} plot for each age.
\item \code{grid.IGE.class}: a heatmap representing the grid. The cells are filled according
to the competition class of each genotype. If \code{age = TRUE}, there will be a \code{grid.DGE} plot for each age.
}
All plots are customizable using resources of the \code{ggplot2} library.
}
}
\description{
This function provides responses extracted from the genetic-spatial
competition model fitted using \code{\link[=asr]{asr()}}.
}
\details{
The genetic-spatial competition model provides the direct (DGE) and indirect
genetic effects (IGE) of each genotype. The DGE represents the "pure" performance
of the genotype, while the IGE is the related to the average effect of the genotype on
the genotypic value of its neighbours. The higher the IGE, the more aggressive is the genotype.
Here, we use the classification proposed by \insertCite{ferreira_novel_2023;textual}{competition}
to define competition classes:

\deqn{\begin{cases} c_i > \overline{c} + sd(c) \rightarrow \text{Aggressive} \\ \overline{c} + sd(c) > c_i > \overline{c} - sd(c) \rightarrow \text{Homeostatic} \\ c_i < \overline{c} - sd(c) \rightarrow \text{Sensitive} \end{cases}}

where \eqn{c_i} is the IGE of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} genotype, \eqn{\overline{c}} is
the mean IGE, and \eqn{sd(c)} is the standard deviation of the IGE. This classification is
detailed in the plot \code{IGE.density}.

The total genotypic value (TGV) is given by:

\deqn{TGV_i =  d_i + CIF \times c_i}

where \eqn{d_i} is the DGE of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} genotype, and \eqn{CIF} is
the overall competition intensity factor, previously computed in the function
\code{\link[=prep]{prep()}}. If \code{weight.tgv = TRUE}, the DGE and IGE will be
multiplied by their respective reliabilities (\eqn{r_{d_i}^2} and \eqn{r_{c_i}^2}):

\deqn{wTGV_i = r_{d_i}^2 \times d_i + r_{c_i}^2 \times {CIF \times c_i}}
}
\examples{
\donttest{
 comp_mat = prep(data = euca, gen = 'clone', repl = 'block', area = 'area', 
                      ind = 'tree', age = 'age', row = 'row', col = 'col', 
                      dist.col = 3, dist.row = 2, trait = 'mai', method = 'SK',
                      n.dec = 3, verbose = TRUE)
 
 model = asr(prep.out = comp_mat, 
                  fixed = mai~ age, 
                  random = ~ block:age, 
                  cor = TRUE, maxit = 50)
                  
  results = resp(prep.out = comp_mat, model = model, weight.tgv = FALSE)
 }
}
\references{
\insertAllCited{}
}
\seealso{
\link{prep}, \link{asr}, \link{ggplot2}
}
