% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{resp}
\alias{resp}
\title{Extract outputs from the genetic-spatial competition model}
\usage{
resp(prep.out, model, weight.tgv = FALSE, sd.class = 1)
}
\arguments{
\item{prep.out}{A \code{comprep} object.}

\item{model}{An \code{asreml} object, preferably obtained using the \code{\link[=asr]{asr()}} function.}

\item{weight.tgv}{A logical value. If \code{TRUE}, the function will use the direct and
indirect genetic effects' reliability as a weight when estimating the total genotypic
value. Defaults to \code{FALSE}. See Details for more information on these methods.}

\item{sd.class}{A numeric value defining the weight to establish competition classes.
Defaults to 1. See Details for more information.}
}
\value{
The function returns:
\itemize{
\item \code{lrt} : a data frame with the Likelihood Ratio Test results.
\item \code{varcomp} : A data frame summarizing the random parameter vector
(object$vparameters). Variance component ratios are included if param = "gamma",
and a measure of precision (standard error) is included along with boundary
constraints at termination and the percentage change in the final iteration.
\item \code{blups} : A list with a dataframe containing the direct (DGE) and indirect genetic effects (IGE), their standard errors,
the competition class of each genotype and the total genotypic value (TGV). If \code{age = TRUE} in
\code{\link[=prep]{prep()}}, \code{blup} will contain a further dataframe the within-ages DGE and IGE. If
other random effects were declared in the model, \code{blup} will contain a last data frame
with their BLUPs.
}
}
\description{
This function provides responses extracted from the genetic-spatial
competition model fitted using \code{\link[=asr]{asr()}}.
}
\details{
The genetic-spatial competition model provides the direct (DGE) and indirect
genetic effects (IGE) of each genotype. The DGE represents the "pure" performance
of the genotype, while the IGE is the related to the average effect of the genotype on
the genotypic value of its neighbours. The higher the IGE, the more aggressive is the genotype.
Here, we use the classification proposed by \insertCite{ferreira_novel_2023;textual}{gencomp}
to define competition classes:

\deqn{\begin{cases} c_i > \overline{c} + sd(c) \times \alpha \rightarrow \text{Aggressive} \\ \overline{c} + sd(c) \times \alpha > c_i > \overline{c} - sd(c) \times \alpha \rightarrow \text{Homeostatic} \\ c_i < \overline{c} - sd(c) \times \alpha \rightarrow \text{Sensitive} \end{cases}}

where \eqn{c_i} is the IGE of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} genotype, \eqn{\overline{c}} is
the mean IGE, \eqn{sd(c)} is the standard deviation of the IGE and
\eqn{\alpha} is a weight (defaults to 1). This classification is
detailed in the plot \code{IGE.density}.

The total genotypic value (TGV) is given by:

\deqn{TGV_i =  d_i + CIF \times c_i}

where \eqn{d_i} is the DGE of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} genotype, and \eqn{CIF} is
the overall competition intensity factor, previously computed in the function
\code{\link[=prep]{prep()}}. If \code{weight.tgv = TRUE}, the DGE and IGE will be
multiplied by their respective reliabilities (\eqn{r_{d_i}^2} and \eqn{r_{c_i}^2}):

\deqn{wTGV_i = r_{d_i}^2 \times d_i + r_{c_i}^2 \times {CIF \times c_i}}
}
\examples{
\donttest{
 library(gencomp)
 comp_mat = prep(data = euca, gen = 'clone', repl = 'block', area = 'area', 
                 ind = 'tree', age = 'age', row = 'row', col = 'col', 
                 dist.col = 3, dist.row = 2, trait = 'mai', method = 'SK',
                 n.dec = 3, verbose = TRUE)
 model = asr(prep.out = comp_mat, 
             fixed = mai~ age, 
             random = ~ block:age, 
             cor = TRUE, maxit = 20,
             lrtest = FALSE)
                  
 results = resp(prep.out = comp_mat, model = model, weight.tgv = FALSE)
 
 results$varcomp  # Variance components
 results$blups$main  # BLUPs (DGE, IGE and TGV, main effects)
 results$blups$within  # BLUPs (DGE, IGE and TGV, within-age effects)
 
 # To export these results to the working directory, use:
 # write.csv(results$blups$main, file = "working/directory/filename.csv")
 }
}
\references{
\insertAllCited{}
}
\seealso{
\link{prep}, \link{asr}
}
