% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{summary.comresp}
\alias{summary.comresp}
\title{Summary of the \code{comresp} object}
\usage{
\method{summary}{comresp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{comresp}}

\item{...}{Currently not used.}
}
\value{
The function returns a list with the variance components (\code{varcomp})
and a dataframe informing which genotypes had the highest DGE, the highest and
the lowest IGE, and the highest TGV. This dataframe also informs the number of
aggressive, homeostatic and sensitive. If a multi-age model was fitted, the output
list will have another dataframe with the same information for each age.
}
\description{
A brief summary of the \code{comresp} object.
}
\examples{
\donttest{
library(GenComp)
comp_mat = prep(data = euca, gen = 'clone', repl = 'block', area = 'area', 
                ind = 'tree', age = 'age', row = 'row', col = 'col', 
                dist.col = 3, dist.row = 2, trait = 'mai', method = 'SK',
                n.dec = 3, verbose = TRUE)
 model = asr(prep.out = comp_mat, 
             fixed = mai~ age, 
             random = ~ block:age, 
             cor = TRUE, maxit = 20,
             lrtest = FALSE)
 results = resp(prep.out = comp_mat, model = model, weight.tgv = FALSE)
 
 summary(results)
}

}
\seealso{
\link{resp}
}
